remoção código legado; ajustes pontuais nas entidades; Criação dos repositorios ClientRepository.kt e TransactionRepository.kt integrado ao Spring Data

+++ solutions/devsprint-michel-ferreira-1/src/main/kotlin/com/devpass/challengehexagonal/domain/model/Transaction.kt
@@ -1,9 +1,10 @@
 package com.devpass.challengehexagonal.domain.model
 
+import java.math.BigDecimal
 import java.time.LocalDateTime
 
 data class Transaction(
     val establishment: String,
-    val value: Double,
+    val value: BigDecimal,
     val transactionDate: LocalDateTime = LocalDateTime.now()
 )
